+++ docs/fr_FR/changelog.md
@@ -3,6 +3,17 @@
 >**IMPORTANT**
 S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 25/04/2025 Beta 1.3.5
+- Ajout d'un timeout pour les requêtes
+- Ajout d'un retry pour les requêtes
+- Correction calendrier Booking
+
+# 09/04/2025 Beta 1.3.4
+- Petites corrections
+
+# 05/04/2025 Beta 1.3.3
+- Correction bug ical2calendar sur les évènements j-1 à j+7
+
 # 06/03/2025 Beta 1.2.8
 - Correction bug ical2calendar
 - Ajout des commandes j-1 ainsi que j+2 à j+7